Ora2ES: more accurate handling of case when no new data found.
Sending empty file to ES produces a null_pointer_exception error.

+++ Utils/Elasticsearch/process.sh
@@ -42,7 +42,11 @@ putDataToES() {
 }
 
 getDataFromOracle > $ORACLE_OUT 
-convertDataToESFormat < $ORACLE_OUT >$ES_IN 
-putDataToES <$ES_IN
-echo OFFSET="'${NEW_OFFSET}'" > "$OFFSET_FILE"
+if [ -s "$ORACLE_OUT" ]; then
+  convertDataToESFormat < $ORACLE_OUT >$ES_IN
+  putDataToES <$ES_IN
+  echo OFFSET="'${NEW_OFFSET}'" > "$OFFSET_FILE"
+else
+  log "No data changed since last offset."
+fi
 rm -f $ORACLE_OUT $ES_IN
